refactor: 도메인별 enum 패키지 정리
