+++ src/main/java/chess/domain/plate/Direction.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.plate;
 
 import lombok.Getter;
