+++ src/test/java/chess/domain/piece/KingTest.java
@@ -1,5 +1,8 @@
 package chess.domain.piece;
 
+import chess.domain.plate.File;
+import chess.domain.plate.Rank;
+import chess.domain.team.Team;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.DisplayName;
