+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -1,10 +1,12 @@
 package chess.domain.piece;
 
+import chess.domain.plate.File;
+import chess.domain.plate.Rank;
+import chess.domain.team.Team;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class KnightTest {
     @Test
