+++ src/test/java/chess/domain/plate/ChessPlateTest.java
@@ -1,9 +1,7 @@
 package chess.domain.plate;
 
 import chess.domain.MovingDirection;
-import chess.domain.piece.File;
 import chess.domain.piece.PiecePosition;
-import chess.domain.piece.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
