+++ src/test/java/chess/service/ChessServiceTest.java
@@ -1,6 +1,7 @@
 package chess.service;
 
-import chess.domain.piece.*;
+import chess.domain.plate.File;
+import chess.domain.team.Team;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
